Move the bookHistos() method to the init of the class
This change should be performed in every class, otherwise
there are problems whenever a cluster job is sent.
The current called to the bookHistos is placed in the
processRunHeader, which is not called whenever events
are skipped. This is a failure in the design of the classes
(probably in the LCIO classes, given that a cluster job
cannot access to the run header...)