Include noise information for the alibava clusters
This change has to be checked in the downstream
processing. In principle this change does not
interfere with any other step as the noise are
included inside the TrackerData of the cluster
which is afterwards propagated as the RawHits.
However, it is needed to check the full processing
chain to be sure