Fix bug when selected chip doesn't contain the first one
The bug is appearing because of the momentaneous
mix between the number of elements to be processed
in the input collection versus the chip number
correspoding to that element. When there is no
chip 0 selected, then the chip-number IS NOT the
index of the element of the input col. The alibava
classes have been scanned searching for the presence
of this bug in other classes but didn't found anything.
Be careful, though.